requirements.txt
allure-pytest-bdd
api-client
black>=20.8b1
boto3
celery[redis]>=5.0.2
dj-database-url
django==3.1
django-dotenv
django-extra-fields
django-filter
django-fsm
django-health-check
django-polymorphic
django-redis
django-rest-polymorphic
django-sequences
django-storages
django-treebeard
django-webpack-loader
django_extensions
djangorestframework
factory-boy
freezegun
git+https://github.com/alphagov/govuk-frontend-jinja.git
gunicorn
jinja2
lxml
moto
parsec
psycopg2-binary
pytest
pytest-bdd
pytest-cov
pytest-django
pytest-responses
requests_oauthlib
responses
sentry-sdk
sphinx

# This needs to be pinned below 1.26.0 for responses to work until
# https://github.com/getsentry/responses/issues/354 is resolved.
urllib3<1.26.0

werkzeug
whitenoise
xlrd
xmldiff
